Make changes requested from Pull Request
This commit will update the code to be in line with the changes requested in the most recent pull request. Most notably there will be new helper methods in place of repititious code, and all HTML-rendering non-async functions moved into a new file named traveler_dashboard_helper. This will help the project to  appear cleaner and much more organized.
