--- expected_output.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width" />
-    <title>repl.it</title>
-    <link href="style.css" rel="stylesheet" type="text/css" />
-  </head>
-  <body>
-    <h1>Welcome, Josh</h1>
-    <section class="upcoming-trips">
-      <h3>Upcoming Trips:</h3>
-      <article class="trip">
-        <p><b>Trip Date:</b> 3/12/21</p>
-        <p><b>Trip Location:</b> Jakarta, Indonesia</p>
-        <p><b>Trip Status:</b> approved</p>
-        <p><b>Trip Duration:</b> 5 days</p>
-        <img src="" alt="" />
-      </article>
-    </section>
-    <section class="past-trips">
-      <h3>Past Trips:</h3>
-      <article class="trip">
-        <p><b>Trip Date:</b> 3/12/19</p>
-        <p><b>Trip Location:</b> Jakarta, Indonesia</p>
-        <p><b>Trip Status:</b> pending</p>
-        <p><b>Trip Duration:</b> 5 days</p>
-        <img src="" alt="" />
-      </article>
-    </section>
-    <section class="total-spending">
-      <summary><b>Total Spent:</b> $12,340</summary>
-    </section>
-    <script src="script.js"></script>
-  </body>
-</html>
