+++ src/travel_tracker_service.js
@@ -65,77 +65,9 @@ async function getTravelerDestinations(travelerTrips, traveler) {
   return { travelerTrips, traveler, travelerDestinations };
 }
 
-// Change this function name to getTravelerDashboardData
-async function getTravelerDashboardData(
-  traveler,
-  travelerDestinations,
-  travelerTrips
-) {
-  let lodgingCosts = [];
-  let i;
-  for (i = 0; i < travelerDestinations.length; i++) {
-    lodgingCosts.push(
-      travelerDestinations[i].estimatedLodgingCostPerDay *
-      travelerTrips[i].duration
-    );
-  }
-
-  let flightCosts = [];
-  for (i = 0; i < travelerDestinations.length; i++) {
-    flightCosts.push(
-      travelerDestinations[i].estimatedFlightCostPerPerson *
-      travelerTrips[i].travelers
-    );
-  }
-
-  let flightsPlusLodging = [];
-  for (i = 0; i < travelerDestinations.length; i++) {
-    flightsPlusLodging.push(lodgingCosts[i] + flightCosts[i]);
-  }
-  var totalAmountCustHasSpentBeforeAgentFee = flightsPlusLodging.reduce(
-    function (a, b) {
-      return a + b;
-    },
-    0
-  );
-  var totalAfterAgentFee = totalAmountCustHasSpentBeforeAgentFee * 1.1;
-
-  return {
-    traveler,
-    travelerTrips,
-    travelerDestinations,
-    lodgingCosts,
-    flightCosts,
-    flightsPlusLodging,
-    totalAmountCustHasSpentBeforeAgentFee,
-    totalAfterAgentFee,
-  };
-}
-
-async function getApprovedTripsForAgent() {
-  const url =
-    "https://fe-apps.herokuapp.com/api/v1/travel-tracker/data/trips/trips";
-  let response = await fetch(url);
-  let tripsResponse = await response.json();
-  let trips = tripsResponse.trips;
-  return { trips };
-}
-
-async function getDestinationsForAgent() {
-  const url =
-    "https://fe-apps.herokuapp.com/api/v1/travel-tracker/data/destinations/destinations";
-  let response = await fetch(url);
-  let destinationsResponse = await response.json();
-  let destinations = destinationsResponse.destinations;
-  return { destinations };
-}
-
 export {
   isValidTraveler,
   getSingleTraveler,
   getTravelerTrips,
   getTravelerDestinations,
-  getTravelerDashboardData,
-  getApprovedTripsForAgent,
-  getDestinationsForAgent,
 };
